examples/mcp-python/mcp-client/pyproject.toml
[project]
name = "mcp-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-genai>=1.10.0",
    "google-generativeai>=0.8.4",
    "mcp[cli]>=1.6.0",
]
